requirements.txt
argh==0.26.2
awscli==1.18.212
boto3==1.16.52
botocore==1.19.52
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
docutils==0.15.2
ecdsa==0.14.1
envs==1.3
ez-setup==0.9
Flask==1.1.2
Flask-Resize==2.0.4
grpcio==1.34.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jwt==1.1.0
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pilkit==2.0
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycognito==0.1.5
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
s3transfer==0.3.4
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
webapp2==2.5.2
Werkzeug==1.0.1
